requirements.txt
streamlit
pandas
numpy
joblib
spotipy
python-dotenv
scikit-learn
